+++ code/external/fftreal/headers/FFTRealFixLenParam.h
@@ -44,7 +44,7 @@ public:
    // Over this bit depth, we use direct calculation for sin/cos
    enum {	      TRIGO_BD_LIMIT	= 12  };
 
-	typedef	float	DataType;
+	typedef	double	DataType;
