+++ code/external/fftreal/sources/FFTRealSelect.cpp
@@ -35,7 +35,7 @@ namespace ffft
 
 
 template <int P>
-float *	FFTRealSelect <P>::sel_bin (float *e_ptr, float *o_ptr)
+double *	FFTRealSelect <P>::sel_bin (double *e_ptr, double *o_ptr)
 {
 	return (o_ptr);
 }
@@ -43,7 +43,7 @@ float *	FFTRealSelect <P>::sel_bin (float *e_ptr, float *o_ptr)
 
 
 template <>
-inline float *	FFTRealSelect <0>::sel_bin (float *e_ptr, float *o_ptr)
+inline double *	FFTRealSelect <0>::sel_bin (double *e_ptr, double *o_ptr)
 {
 	return (e_ptr);
 }
